Added defaults to inspect
This adds defaults to the `Kernal.inspect` function which avoids
truncating information. The assumption here is that it's better to
provide defaults that will *not* result in the loss of information by
default. If this is not suitable then the implementer can override the
options and put it into a default *unsafe* configuration.